-доделаны стейты на экране поиска и экране сохраненных треков -добавление удаление треков

+++ data/src/main/java/kadyshev/dmitry/data/dataSource/TrackDao.kt
@@ -14,12 +14,12 @@ interface TrackDao {
     suspend fun insertTrack(track: TrackDBModel)
 
     @Query("SELECT * FROM tracks WHERE id = :id")
-    suspend fun getTrackById(id: Int): TrackDBModel?
+    suspend fun getTrackById(id: Long): TrackDBModel?
 
     @Query("SELECT * FROM tracks")
     fun getAllTracks(): Flow<List<TrackDBModel>>
 
     @Query("DELETE FROM tracks WHERE id = :id")
-    suspend fun deleteTrack(id: Int)
+    suspend fun deleteTrack(id: Long)
 
 }
